feat: add a new buildspec configuration that runs both unit and integ tests
motivation: rather than running integ and unit tests in separate CodeBuild projects and for PR builds,
consolidate them into a single project which will run the sequentially. This is mostly to simplify the
current infrastructure setup and make PRs less confusing as we will 1 set of build logs for every PR build.

The current PR build runs these separately and posts logs/comments for each separately, which is more tedious.