feat: Adds support for Placeholders in TrainingStep to set S3 location for InputDataConfig and OutputDataConfig (#142)
Currently, it is not possible to specify the output path to TrainingStep. It needs to be defined at the Estimator level, which does not support placeholders. This change will make it possible to insert a placeholder output path from the TrainingStep definition and propagate it dynamically to the Estimator.

This change change also makes the TrainingStep data parameter compatible with placeholders.

There are other feature requests to make other TrainingStep arguments and ProcessingStep arguments compatible with placeholders. They will be addressed in a separate PR where the implementation could maybe be done at a higher level to avoid repetition.

Closes  #98 #97 #80