chore: add retry to SageMaker steps in integration tests (#162)
Currently, we observe a few different failures that occur during integration tests, which get
executed as a part of the PR build as well as pushes to branches.

`createModel` and `createEndpoint` particularly see failures most frequently and they are
primarily:
* Rate exceeded  - `ThrottlingException`.

This change defines a default retry strategy that makes 5 attempts, over an interval of 5
seconds, which backs off with a multiplier of 2. The methodology behind this strategy is
naive and  may need some calibration. It should reduce the frequency of failures in the
short term.

We can adjust the retry strategy as we go and expand to something more API specific as
the need arises.